Fixed some issues
Added contact page, for contacting owner,  added delete modal for posts, fixed bug with cloning of posts, Logind Logout fixed on index page when user is loged showin him logout and admin link in navigation bar, showin all posts to admin on home page, and showing onli published pages to regular user

+++ admin/includes/view_all_users.php
@@ -79,10 +79,14 @@
 
 
 
-    if(isset($_GET['delete'])){
-        $the_user_id = $_GET['delete'];
-        $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-        $result = mysqli_query($connection,$query);
-        header("Location: users.php");
+    if(isset($_GET['delete'])){  
+        if(isset($_SESSION['user_rile'])){
+            if($_SESSION['user_rile'] == 'admin'){
+                $the_user_id = mysqli_real_escape_string($connection,$_GET['delete']);
+                $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+                $result = mysqli_query($connection,$query);
+                header("Location: users.php");
+            }
+        }
     }
 ?>
